Slightly change the way the normalize filter works: It was still possible to generate a clipping signal, because peak_amplitude was being calculated from amplitude which had been passed through the IIR filters. Now peak_amplitude is correctly calculated from the input.
Also, since it is no longer possible to generate a clipping signal (the max-gain
is limited by the input signal), the maximum amount we can increase the
loudness of the audio is quite limited. (Tests indicate we can expect at most a gain of 1.3x)
To combat this the reference level has been dropped by 11dB. This means all songs will be
levelled to a quieter level, and quiet songs do not need to be boosted by much (if at all).

The only downside is that the maximum output volume is also quite a bit lower. (Use an external amplifier / speakers!)